fix folder

+++ app.py
@@ -13,7 +13,8 @@ def download_slide():
     url = request.args.get("url-link")
     caminho_slides = os.getcwd() + os.sep + 'slides'
     for the_file in os.listdir(caminho_slides):
-        os.unlink(caminho_slides + os.sep + the_file)
+        if (the_file is not "index.txt"):
+            os.unlink(caminho_slides + os.sep + the_file)
     
     name = gospel.createSlide(url)
     return send_file(caminho_slides + os.sep + name,  as_attachment=True, attachment_filename = name)
